+++ security/src/main/java/com/networknt/security/JwtVerifyHandler.java
@@ -16,18 +16,21 @@
 
 package com.networknt.security;
 
+import com.networknt.config.Config;
+import com.networknt.status.Status;
 import com.networknt.utility.Constants;
+import com.networknt.utility.ExpiredTokenException;
 import io.undertow.Handlers;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.HttpServerExchange;
-import io.undertow.server.handlers.ResponseCodeHandler;
 import io.undertow.util.HeaderMap;
 import io.undertow.util.Headers;
 import io.undertow.util.HttpString;
 import io.undertow.util.StatusCodes;
 import org.jose4j.jwt.JwtClaims;
-import org.slf4j.ext.XLogger;
-import org.slf4j.ext.XLoggerFactory;
+import org.jose4j.jwt.consumer.InvalidJwtException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.regex.Pattern;
 
@@ -35,7 +38,12 @@ import java.util.regex.Pattern;
  * Created by steve on 01/09/16.
  */
 public class JwtVerifyHandler implements HttpHandler {
-    static final XLogger logger = XLoggerFactory.getXLogger(JwtVerifyHandler.class);
+    static final Logger logger = LoggerFactory.getLogger(JwtVerifyHandler.class);
+
+    static final String STATUS_INVALID_JWT_TOKEN = "ERR1000";
+    static final String STATUS_JWT_TOKEN_EXPIRED = "ERR1001";
+    static final String STATUS_MISSING_JWT_TOKEN = "ERR1002";
+
 
     private volatile HttpHandler next;
 
@@ -55,16 +63,24 @@ public class JwtVerifyHandler implements HttpHandler {
                 headerMap.add(new HttpString(Constants.CLIENT_ID), claims.getStringClaimValue(Constants.CLIENT_ID));
                 headerMap.add(new HttpString(Constants.USER_ID), claims.getStringClaimValue(Constants.USER_ID));
                 headerMap.add(new HttpString(Constants.SCOPE), claims.getStringListClaimValue(Constants.SCOPE).toString());
+
+
                 next.handleRequest(exchange);
-            } catch (Exception e) {
+            } catch (InvalidJwtException e) {
                 // only log it and unauthorized is returned.
                 logger.error("Exception: ", e);
-                exchange.setStatusCode(StatusCodes.UNAUTHORIZED);
-                exchange.endExchange();
+                Status status = new Status(STATUS_INVALID_JWT_TOKEN);
+                exchange.setStatusCode(status.getStatusCode());
+                exchange.getResponseSender().send(Config.getInstance().getMapper().writeValueAsString(status));
+            } catch (ExpiredTokenException e) {
+                Status status = new Status(STATUS_JWT_TOKEN_EXPIRED);
+                exchange.setStatusCode(status.getStatusCode());
+                exchange.getResponseSender().send(Config.getInstance().getMapper().writeValueAsString(status));
             }
         } else {
-            exchange.setStatusCode(StatusCodes.UNAUTHORIZED);
-            exchange.endExchange();
+            Status status = new Status(STATUS_MISSING_JWT_TOKEN);
+            exchange.setStatusCode(status.getStatusCode());
+            exchange.getResponseSender().send(Config.getInstance().getMapper().writeValueAsString(status));
         }
     }
 
@@ -77,6 +93,4 @@ public class JwtVerifyHandler implements HttpHandler {
         this.next = next;
         return this;
     }
-
-
 }